[Fix]: CalendarUiState의 날짜 타입을 LocalDate로 수정
